Subscription documentation and refactoring
